minor code cleanup and documentation updates in core.rs, move core.rs tests to parse.rs
